test: split e2e suite into per-topic module files
Break tests/e2e/tests/lib.rs (3300 lines) into topic-scoped modules:
  cnight (lifecycle), governance (council / tech-auth / d-param /
  ariadne), rpc_abuse (ddos / replay), contract_state, and operational.
  All modules compile into one e2e_tests binary so the global faucet
  manager, pre-deploy gate, and deploy serial mutex remain shared
  across the run -- splitting into sibling test binaries would have
  duplicated those statics per binary and broken the gates.

  lib.rs now holds only the shared statics + three helpers
  (finished_pre_deploy_test, wait_before_deploying,
global_faucet_manager,
  now pub(crate)) plus `mod` declarations. Per-module runs are possible
  via the module-prefix filter, e.g. `cargo test ... cnight::`.

  README.md updated with the new layout, the cargo-test filter syntax,
  and a note that `cargo check --tests` skips the integration target
  because [[test]] has test = false.
